Refactor comments and codes for the browser internal components.
